[f] Add fixed node types and datatype names
Fixed nodes are similar to plain old variables whose values are changed
manually through a setter. Nothing special happens in the 'run' function
of fixed nodes.

Added a template customization point for assigning pretty names to certain
datatypes. Newly introduced datatypes should also overload
rsp::DataTypeName and provide a name to be used by the rest of the
system for serialization and UI. By default, the name provided by typeid
is used, which is not consistent accross platforms.