[f] Redesign of the rsp::base lib
Reduced the port and node hierarchies to 3, respectively one classes.
Removed event ports and observer flags, as they felt unnecessary.
Alternatives will have to be found later. Added a sentinel type. For now
it's used just to detect cycles in the nodegraph. In the future other
sentinel types may be used to, for example, collect information about
graph execution, such as contention, average node cost etc.

Updated the gui libs as far as possible to reflect the backend changes,
however, they don't yet work as before. Some changes need to be made to
widgets, and some alternatives need to be found to observer flags. I
want to review some gui code anyway, so a big gui refactoring is coming
up.