[m] Add build system enhancements and fixes
Started using cmake-presets.json and thus bumped the minimum required
cmake version to 3.20. It is now way more comfortable to switch between
common configurations in vscode at least. Each configuration preset
gets its own build folder so you can for e.g. switch back and forth between
 msvc and clang on windows to test out different warnings without
 needing to do a full rebuild each time.

Added a custom target that copies the compilation database to the folder
"build". This is the folder where clangd looks for it by default, so
users of clangd + clang/gcc should get a good experience out of the box
now. There are also presets for each of the sanitizers.

Disabled static analysis by default. The around 10-20x increase
in build times was not worth it to have them constantly on. When
iterating, build speed is really important, and most IDEs enable you to
 run clang-tidy automatically on the current file anyway. Static
 analysis with clang-tidy, cppcheck and include-what-you-use can be
 reenabled with a configuration switch at any time. There is also
 a special configuration preset that only enables this flag.

 Disabled automatic conan install handling for users who want to either
 run conan themselves, or can provide the libraries by other means
 (the build script uses Find library commands anyway, conan is entirely
 optional). The option is enabled for all configuration presets though.