[m] remove copy compile commands cmake target
The copy compile commands was only needed when using clangd together
with build folders other than build.

The CMake tools extension for vscode can apparently copy the compile
commands automatically to the build folder, where clangd looks for them
normally.

Since this was only needed for a specific tooling use case and the tools
can now handle this themselves, it makes no sense to keep it.