[f] add widget data accessors
widget access to data is abstracted away using so called
data accessors. viewers use a data_reader to read the data
and editors use a data_writer to read/write data.

behind the scenes, the data_reader and the data_writer can
use either a pointer to the data directly (equivalent functionality
to how it worked before), or a getter callable.

the data_writer can store a no argument callable to call
when the data is modified, a callable that takes the new data
as an argument, or no callable at all.

this abstraction enables more flexibility in binding widgets
to data, and also simplifies widget factory creation, since all
the variation is handled by the respective data accessor type