[f] add support for composite widgets
implement composite viewers and editors that can be configured at runtime
to represent arbitrary data composed of types for which widgets already
exist. after creating a composed viewer/editor you can add arbitrarily
many sub viewers/editors to it, by specifying how to get to the sub data.
this can be as simple as a member pointer, for basic structs, or
as complex as custom getters and setters.

this should greatly reduce the barrier to writing widgets, since
simple, composed data types do not really need custom widgets. this
also encourages composition, and keeping the amount of custom
widgets to a minimum.