[f] profiler enhancements
simplify profiler code, and change API to measure a starting/ending
pair of timepoints, instead of a fixed frame.

enhance profiler widget. make it an editor instead of viewer, with
ability to change number of profiler samples, as well as (de)activate
it.

add some extra profilers in a separate panel, default deactivated