[f] force based graph layout solver
implement and activate a layout solver component for graph viewers
and editors. the solver tries to space nodes apart from each other
in order to avoid overlapping, and it draws nodes closer together
when connected to each other. the mouse cursor creates a time slowing
effect, so that nodes do not move from under the mouse when the user
is changing connections for example, or wants to force move nodes
around. all layout paramaters can be tweaked through the widget
settins.
note: this single squashed commit has been exactly 1 year in the
making. this was hard, and there is definitely room for optimisations,
but at least the interaction now feels good and useful