[f] add support for vector widgets
because of type erasure, any widget that works generically with a
container, will need to be manually registered in the widget_factory
by registering the container instanced with every known type.
add std::vector in this commit. this will cover 90% of use cases
anyways. for very complex custom types, a specialized widget
can always be registered