[f] exception handling
add ability for nodes to be in an error state. make algorithm nodes
catch any exception thrown by the underlying algorithm's update function
and store it as an error message. additionally, add a 'faulty' flag
to ports, to propagate such error states to downstream nodes.

a nodes error state gets cleared whenever a push event occurs on the
node, since that implies a change in its dependencies, which might lead
to a succesful update. push events clear downstream port faultiness
as well.

add support for highlighting these states in the gui through red and
white blinking of the node that is the origin of an error, and any
faulty ports and links. the node where the exception was thrown also
shows the exception message undeneath